CARDS-1347: Evaluate computed fields serverside Moved editor classes to the `internal` subpackage
