+++ modules/data-entry/src/main/java/io/uhndata/cards/internal/ComputedAnswersEditor.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.uhndata.cards;
+package io.uhndata.cards.internal;
 
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
