+++ modules/data-entry/src/main/java/io/uhndata/cards/internal/ComputedAnswersEditorProvider.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.uhndata.cards;
+package io.uhndata.cards.internal;
 
 import org.apache.jackrabbit.oak.api.CommitFailedException;
 import org.apache.jackrabbit.oak.spi.commit.CommitInfo;
